requirements.txt
pandas==2.2.0
langchain==0.1.7
streamlit==1.31.1
langchain-openai==0.0.6
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary==2.9.9
docarray==0.40.0
python-dotenv==0.19.0